+++ src/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ### {{version_pkg_lupo}} ({{creationDate_changelog}})
+  - package installer removes old sql migration files
+
+### 4.0.1 (October 18, 2023)
   - joomla 4.4 compatibility
   - bugfix reservation button: escape toyname properly